ai_part/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
python-multipart==0.0.12
requests==2.32.3
python-dotenv==1.0.1
openai-whisper

